Changing RunDaily and RunWeekly to use the universe  of dates to decide whether it is a new day or new week. This allows the algo to look ahead at the universe of dates to decide if it is the last date in the period.
Note: If the algo does not run on the specific date in the period, then it will not run in that period. For example, if it is a RunWeekly algo that is set to run in the beginning of the week but it misses that Monday because another algo returned False before it, then it will not run for that week.