For all algos that previously took Dataframe arguments that were dependent on the security universe and the timeline, allow for a string argument that will allow the algo to retrieve the data from the target. The data frames are passed to the target at strategy setup time, and retrieved using the new get_data function, which exposes the previously private _setup_kwargs.
This allows for complete de-coupling of the strategy from the data, and
allows strategies using these algos to be re-used for backtesting
on different date ranges and/or security universes.

For WeighTarget, SelectWhere and SetStat, data frames can still be passed
to the algo for backwards compatibility with existing code. However, string
passing is the preferred way to achieve de-coupling.

For more recently added algos (related to fixed income changes), the string
name *must* be used. The example fixed_income.ipynb has been updated accordingly.