Adding tracking of holding costs to CouponPayingSecurity
This change allows for time varying, instrument-specific, asymmetric holding
costs on CouponPayingSecurity. The cost is based on the absolute value of position,
and can be different depending on whether one is long (i.e. funding) or short (i.e. repo).
This is tracked separately from coupons, which are symmetric, and proportional to
position (rather than absolute position).

The data is optionally passed in to setup under the arguments "cost_long" and "cost_short".